requirements.txt
requests
pandas
tqdm
typer[all]
rich
biopython
